+++ ydb/core/tx/schemeshard/schemeshard_export_scheme_uploader.cpp
@@ -1,5 +1,5 @@
 #include "schemeshard.h"
-#include "schemeshard_export_scheme_uploader.h"
+#include "schemeshard_export_uploaders.h"
 
 #include <ydb/core/backup/common/encryption.h>
 #include <ydb/core/protos/flat_scheme_op.pb.h>
